backend/pyproject.toml
[project]
name = "oneflow-backend"
version = "0.1.0"
description = "OneFlow ERP Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "sqlmodel>=0.0.22",
    "alembic>=1.14.0",
    "argon2-cffi>=23.1.0",
    "pyjwt>=2.9.0",
    "pydantic-settings>=2.7.0",
    "psycopg2-binary>=2.9.10",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.20",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
